+++ spec/components/subscription_links_spec.rb
@@ -32,11 +32,6 @@ describe "subscription links", type: :view do
     assert_select '.gem-c-subscription-links.govuk-\!-margin-bottom-7'
   end
 
-  it "defaults to the initial bottom margin if an incorrect value is passed" do
-    render_component(email_signup_link: "email-signup", feed_link: "singapore.atom", margin_bottom: 20)
-    assert_select "[class^='govuk-\!-margin-bottom-']", false
-  end
-
   it "has no margin class added by default" do
     render_component(email_signup_link: "email-signup", feed_link: "singapore.atom")
     assert_select "[class^='govuk-\!-margin-bottom-']", false